Move collect-metrics lambda to cli tool
