--- project.json
@@ -1,5 +1,0 @@
-{
-  "name": "buildkite-cloudwatch-metrics-publisher",
-  "description": "A publisher for buildkite metrics",
-  "nameTemplate": "buildkite-{{.Function.Name}}"
-}
